requirements.txt
pillow>=10.0.0
pystray>=0.19.4
psutil>=5.9.0
